[ wip ]: improvements
