+++ packages/packages/libs/editor-controls/src/index.ts
@@ -99,3 +99,4 @@ export {
 export { useSyncExternalState } from './hooks/use-sync-external-state';
 export { useElementCanHaveChildren } from './hooks/use-element-can-have-children';
 export { useFontFamilies } from './hooks/use-font-families';
+export { useTypingBuffer, type UseTypingBufferOptions } from './hooks/use-typing-buffer';
